Add theme picker
- separate theme control from selectInput for usability
