+++ app/dashboard/page.js
@@ -1,11 +1,14 @@
+'use client'
 import AreaChart from '@/components/charts/areabar';
 import LineChart from '@/components/charts/linechart';
 import PieChart from '@/components/charts/piechart';
 import VerticalBar from '@/components/charts/verticalbar';
 import { DashCard } from '@/components/dashcard/card'
+import { useState } from 'react';
 
 // Dashboard index page
 export default function Home() {
+ // const [resizeLoad, setResizeLoad ] = useState(false)
   return (
     <main className='h-full overflow-y-auto ' >
       <section className='grid md:grid-cols-2 xl:grid-cols-4  gap-6 my-3 ' >
@@ -43,7 +46,7 @@ export default function Home() {
       </section>
 
       <section className='grid py-3 gap-3 md:grid-cols-2 h-fit '>
-        <div className='p-2 w-full h-fit shadow-xl bg-white hover:scale-[1.01] duration-500'>
+        <div  className='p-2 w-full h-fit shadow-xl bg-white hover:scale-[1.01] duration-500'>
            <AreaChart />
         </div>
          <div className='p-2 w-full h-fit shadow-xl bg-white hover:scale-[1.01] duration-500'>
